Voronoi 2D: factor out main fucntion into voronoi.py.
